thesis/websites/requirements.txt
flask
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
flask-login
